Comments applied
