+++ public/styles/globals.css
@@ -4,8 +4,6 @@
 
 body {
   font-family: Arial, Helvetica, sans-serif;
-  overflow-x: hidden;
-  overflow-y: hidden;
 }
 
 @layer base {
